pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling>=1.25" ]

[project]
name = "openhands"
version = "1.0.6"
description = "OpenHands CLI - Terminal User Interface for OpenHands AI Agent"
readme = "README.md"
license = { text = "MIT" }
authors = [ { name = "OpenHands Team", email = "contact@all-hands.dev" } ]
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
# Using Git URLs for dependencies so installs from PyPI pull from GitHub
# TODO: pin package versions once agent-sdk has published PyPI packages
dependencies = [
  "openhands-sdk==1.1",
  "openhands-tools==1.1",
  "prompt-toolkit>=3",
  "typer>=0.17.4",
  "agent-client-protocol>=0.6.3",
  "pydantic>=2.7",
]

scripts = { openhands = "openhands_cli.simple_main:main", "openhands-acp" = "openhands_acp.main:main" }

[dependency-groups]
# Hatchling wheel target: include the package directory
dev = [
  "flake8>=6",
  "gevent>=24.2.1,<26",
  "pre-commit>=4.3",
  "pyinstaller>=6.15",
  "pytest>=8.4.1",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=5.0.0",
  "pytest-forked>=1.6.0",
  "pytest-xdist>=3.6.1",
  "ruff>=0.12.10",
  "pycodestyle>=2.12.0",
  "pyright[nodejs]>=1.1.405",
  "pre-commit>=4.3.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [ "openhands_cli", "openhands_acp" ]

# uv source pins for internal packages

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes (includes F841: unused-variable)
    "I",    # isort
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
]
# Enforce rules that catch mutable defaults and related pitfalls
# - B006: mutable-argument-default
# - B008: function-call-in-default-argument
# - B039: mutable-contextvar-default
# - RUF012: mutable-class-default
extend-select = ["B006", "B008", "B039", "RUF012"]

[tool.ruff.lint.per-file-ignores]
# Test files often have unused arguments (fixtures, mocks, interface implementations)
"tests/**/*.py" = ["ARG"]




[tool.ruff.lint.isort]
known-first-party = ["openhands"]
combine-as-imports = true
force-single-line = false
lines-after-imports = 2

# Pytest configuration
[tool.pytest.ini_options]
testpaths = [
    "tests"
]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
relative_files = true
omit = [ "tests/*", "**/test_*" ]

[tool.coverage.paths]
source = [
  "openhands_cli/",
  "openhands-cli/openhands_cli/",
]

[tool.pyright]
include = [
    "openhands_cli",
    "openhands_acp",
    "tests",
    "hooks"
]
extraPaths = [
    "openhands_cli",
    "openhands_acp"
]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
useLibraryCodeForTypes = true
typeCheckingMode = "standard"

# UNCOMMENT TO USE EXACT COMMIT FROM AGENT-SDK

# [tool.uv.sources]
# openhands-sdk = { git = "https://github.com/OpenHands/agent-sdk.git", subdirectory = "openhands-sdk", rev = "7b695dc519084e75c482b34473e714845d6cef92" }
# openhands-tools = { git = "https://github.com/OpenHands/agent-sdk.git", subdirectory = "openhands-tools", rev = "7b695dc519084e75c482b34473e714845d6cef92" }
